MxServer/requirements.txt
aiosmtpd==1.4.2
atpublic==3.1.1
attrs==22.1.0
certifi==2022.9.24
PyYAML==6.0
